Create cmdutils package
